+++ src/App.js
@@ -1,35 +1,24 @@
 import React from 'react';
 import './App.css';
-import Inputs from './components/inputs/Inputs';
 import "bootstrap/dist/css/bootstrap.min.css";
 import "bootstrap/dist/js/bootstrap.bundle.min";
 import { library } from '@fortawesome/fontawesome-svg-core'
-import Balance from './components/Balance/Balance';
-import { useEffect, useState } from 'react';
+import { ThemeProvider } from './components/themeProvider/ThemeContext';
+import MainWindow from './components/mainWindow/MainWindow';
 
 
 library.add();
 
 function App() {
 
-  const [incomesList, setIncomesList] = useState([]);
-  const [expensesList, setExpensesList] = useState([]);
-
-  useEffect(() => {
-    const storedExpensesList = JSON.parse(localStorage.getItem("expensesList")) || [];
-    const storedIncomesList = JSON.parse(localStorage.getItem("incomesList")) || [];
-
-      setExpensesList(storedExpensesList);
-
-      setIncomesList(storedIncomesList);
-  }, []);
+  
 
   return (
+    <ThemeProvider>
     <div className="App">
-      <Balance incomesList={incomesList} expensesList={expensesList}/>
-      <Inputs incomesList={incomesList} setIncomesList={setIncomesList} expensesList={expensesList} setExpensesList={setExpensesList}/>
+      <MainWindow/>
     </div>
-  // </ThemeProvider>
+  </ThemeProvider>
 );
 }
