+++ src/components/mainWindow/MainWindow.js
@@ -2,8 +2,9 @@ import { ThemeContext } from "../themeProvider/ThemeContext";
 import Header from "../header/Header";
 import Inputs from "../inputs/Inputs";
 import ChartWrapper from "../chartWrapper/ChartWrapper";
-import { useContext } from "react";
+import { useContext, useEffect, useState } from "react";
 import "./MainWindow.css"
+import Balance from "../Balance/Balance";
 
 const MainWindow = () => {
   const { theme } = useContext(ThemeContext);
@@ -40,11 +41,25 @@ dark: {
 }
   }
   
+  const [incomesList, setIncomesList] = useState([]);
+  const [expensesList, setExpensesList] = useState([]);
+
+  useEffect(() => {
+    const storedExpensesList = JSON.parse(localStorage.getItem("expensesList")) || [];
+    const storedIncomesList = JSON.parse(localStorage.getItem("incomesList")) || [];
+
+      setExpensesList(storedExpensesList);
+
+      setIncomesList(storedIncomesList);
+  }, []);
+
   return (
     <div className="MainWindow" style={{ ...GlobalStyles[theme] }}>
       <Header style={{ ...headerStyles[theme] }}/>
       <ChartWrapper />
-      <Inputs />
+    
+      <Balance incomesList={incomesList} expensesList={expensesList}/>
+      <Inputs incomesList={incomesList} setIncomesList={setIncomesList} expensesList={expensesList} setExpensesList={setExpensesList}/>
     </div>
   );
 };
